Add just - omniversal package manager wrapper
I've had this idea for a while. It's so annoying having to remember
exactly how to do common tasks on the various distros I use, and it
seems much smarter to just automate it away. This was only about a
half-hour's work, and it saves me having to look up how to do this
shit in future :D